tts

Struct Tts

pub struct Tts(/* private fields */);

Implementations

impl Tts

pub fn new(backend: Backends) -> Result<Tts, Error>

Create a new TTS instance with the specified backend.

pub fn default() -> Result<Tts, Error>

pub fn supported_features(&self) -> Features

Returns the features supported by this TTS engine

pub fn speak<S: Into<String>>( &mut self, text: S, interrupt: bool, ) -> Result<Option<UtteranceId>, Error>

Speaks the specified text, optionally interrupting current speech.

pub fn stop(&mut self) -> Result<&Self, Error>

Stops current speech.

pub fn min_rate(&self) -> f32

Returns the minimum rate for this speech synthesizer.

pub fn max_rate(&self) -> f32

Returns the maximum rate for this speech synthesizer.

pub fn normal_rate(&self) -> f32

Returns the normal rate for this speech synthesizer.

pub fn get_rate(&self) -> Result<f32, Error>

Gets the current speech rate.

pub fn set_rate(&mut self, rate: f32) -> Result<&Self, Error>

Sets the desired speech rate.

pub fn min_pitch(&self) -> f32

Returns the minimum pitch for this speech synthesizer.

pub fn max_pitch(&self) -> f32

Returns the maximum pitch for this speech synthesizer.

pub fn normal_pitch(&self) -> f32

Returns the normal pitch for this speech synthesizer.

pub fn get_pitch(&self) -> Result<f32, Error>

Gets the current speech pitch.

pub fn set_pitch(&mut self, pitch: f32) -> Result<&Self, Error>

Sets the desired speech pitch.

pub fn min_volume(&self) -> f32

Returns the minimum volume for this speech synthesizer.

pub fn max_volume(&self) -> f32

Returns the maximum volume for this speech synthesizer.

pub fn normal_volume(&self) -> f32

Returns the normal volume for this speech synthesizer.

pub fn get_volume(&self) -> Result<f32, Error>

Gets the current speech volume.

pub fn set_volume(&mut self, volume: f32) -> Result<&Self, Error>

Sets the desired speech volume.

pub fn is_speaking(&self) -> Result<bool, Error>

Returns whether this speech synthesizer is speaking.

pub fn voices(&self) -> Result<Vec<Voice>, Error>

Returns list of available voices.

pub fn voice(&self) -> Result<Option<Voice>, Error>

Return the current speaking voice.

pub fn set_voice(&mut self, voice: &Voice) -> Result<(), Error>

Set speaking voice.

pub fn on_utterance_begin( &self, callback: Option<Box<dyn FnMut(UtteranceId)>>, ) -> Result<(), Error>

Called when this speech synthesizer begins speaking an utterance.

pub fn on_utterance_end( &self, callback: Option<Box<dyn FnMut(UtteranceId)>>, ) -> Result<(), Error>

Called when this speech synthesizer finishes speaking an utterance.

pub fn on_utterance_stop( &self, callback: Option<Box<dyn FnMut(UtteranceId)>>, ) -> Result<(), Error>

Called when this speech synthesizer is stopped and still has utterances in its queue.

pub fn screen_reader_available() -> bool